Rename h5_conversion.py as keras_h5_conversion.py
