+++ python/tensorflowjs/converters/BUILD
@@ -1,6 +1,6 @@
 py_library(
-    name = "h5_conversion",
-    srcs = ["h5_conversion.py"],
+    name = "keras_h5_conversion",
+    srcs = ["keras_h5_conversion.py"],
     deps = [
         "//tensorflowjs:expect_h5py_installed",
         "//tensorflowjs:expect_keras_installed",
@@ -10,11 +10,11 @@ py_library(
 )
 
 py_test(
-    name = "h5_conversion_test",
-    srcs = ["h5_conversion_test.py"],
+    name = "keras_h5_conversion_test",
+    srcs = ["keras_h5_conversion_test.py"],
     srcs_version = "PY2AND3",
     deps = [
-        ":h5_conversion",
+        ":keras_h5_conversion",
         "//tensorflowjs:expect_h5py_installed",
         "//tensorflowjs:expect_keras_installed",
         "//tensorflowjs:expect_numpy_installed",
@@ -26,7 +26,7 @@ py_binary(
     srcs = ["converter.py"],
     srcs_version = "PY2AND3",
     deps = [
-        ":h5_conversion",
+        ":keras_h5_conversion",
         "//tensorflowjs:expect_h5py_installed",
     ],
 )
